feat: Add an authentication contract and logging
Each DB Table would hosts a contract `auditLog.sol`. All the DB
operations would be saved in the event log in this `auditLog.sol`.

`authRole.sol` is a contract for authenticating the permission of all
the dropkit service requester. At the same time, `authRole.sol` is the
only one (both contract account and EOA) which can interact with the
contract maintained by each Table.